feat: support gzip compression
